requirements.txt
# Backend API Requirements for Render.com Deployment
# Face Recognition Attendance System

# Core Framework
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Face Recognition & Computer Vision
face_recognition==1.3.0
opencv-python-headless==4.9.0.80
numpy==1.24.3
pillow==10.1.0

# Database
mysql-connector-python==9.3.0
SQLAlchemy==2.0.41
Flask-SQLAlchemy==3.1.1

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# Environment & Security
python-dotenv==1.0.0
Werkzeug==3.0.1

# Development (optional for local testing)
# pytest==7.4.3
# python-multipart==0.0.6

